+++ places/views.py
@@ -26,22 +26,27 @@ class PlaceTypeViewSet(viewsets.ViewSet):
 
     @staticmethod
     def create(request):
-        serializer = CreatePlaceTypeSerializer(data=request.data)
+        user_role = request.user.user_role
+        if user_role == 'superuser' or user_role == 'admin':
+            serializer = CreatePlaceTypeSerializer(data=request.data)
 
-        if not serializer.is_valid():
-            return Response({
-                'message': 'Some fields are missing',
-                'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
-        data = serializer.validated_data
+            if not serializer.is_valid():
+                return Response({
+                    'message': 'Some fields are missing',
+                    'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+            data = serializer.validated_data
 
-        try:
-            place_type = PlaceTypes.objects.create(place_type=data['place_type'])
-        except IntegrityError:
-            return Response({
-                'message': 'Place Type already exists.'
-            }, status=status.HTTP_400_BAD_REQUEST)
+            try:
+                place_type = PlaceTypes.objects.create(place_type=data['place_type'])
+            except IntegrityError:
+                return Response({
+                    'message': 'Place Type already exists.'
+                }, status=status.HTTP_400_BAD_REQUEST)
 
-        return Response(data=PlaceTypeSerializer(place_type).data, status=status.HTTP_201_CREATED)
+            return Response(data=PlaceTypeSerializer(place_type).data, status=status.HTTP_201_CREATED)
+        else:
+            response = 'You are not allowed to create this place type.'
+            return Response(data=response, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     def update(request, pk=None):
@@ -90,22 +95,28 @@ class PriceCategoriesViewSet(viewsets.ViewSet):
 
     @staticmethod
     def create(request):
-        serializer = CreatePriceCategorySerializer(data=request.data)
+        user_role = request.user.user_role
 
-        if not serializer.is_valid():
-            return Response({
-                'message': 'Some fields are missing',
-                'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
-        data = serializer.validated_data
+        if user_role == 'superuser' or user_role == 'admin':
+            serializer = CreatePriceCategorySerializer(data=request.data)
 
-        try:
-            price_category = PriceCategories.objects.create(category_name=data['category_name'])
-        except IntegrityError:
-            return Response({
-                'message': 'Price Category already exists.'
-            }, status=status.HTTP_400_BAD_REQUEST)
+            if not serializer.is_valid():
+                return Response({
+                    'message': 'Some fields are missing',
+                    'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+            data = serializer.validated_data
 
-        return Response(data=PriceCategorySerializer(price_category).data, status=status.HTTP_201_CREATED)
+            try:
+                price_category = PriceCategories.objects.create(category_name=data['category_name'])
+            except IntegrityError:
+                return Response({
+                    'message': 'Price Category already exists.'
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+            return Response(data=PriceCategorySerializer(price_category).data, status=status.HTTP_201_CREATED)
+        else:
+            response = 'You are not allowed to create this price category.'
+            return Response(data=response, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     def update(request, pk=None):
@@ -149,50 +160,55 @@ class PlacesViewSet(viewsets.ViewSet):
     @staticmethod
     def create(request):
         user_id = request.user.id
-        serializer = CreatePlaceSerializer(data=request.data)
+        user_email_verified = request.user.email_confirmed
 
-        if not serializer.is_valid():
-            return Response({
-                'message': 'Some fields are missing',
-                'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+        if user_email_verified:
+            serializer = CreatePlaceSerializer(data=request.data)
 
-        data = serializer.validated_data
-        audio_file = request.data['audio_file']
-        audio_file_name = ''
+            if not serializer.is_valid():
+                return Response({
+                    'message': 'Some fields are missing',
+                    'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
-        if audio_file:
-            audio_file_name = f"{user_id}_{time.time()}_audio.wav"
-            bucket_name = os.getenv('AWS_PLACE_AUDIO_BUCKET_NAME')
-            upload = upload_file_to_aws(file_name=audio_file_name, bucket=bucket_name)
+            data = serializer.validated_data
+            audio_file = request.data['audio_file']
+            audio_file_name = ''
 
-            if upload is True:
-                os.remove(audio_file_name)
+            if audio_file:
+                audio_file_name = f"{user_id}_{time.time()}_audio.wav"
+                bucket_name = os.getenv('AWS_PLACE_AUDIO_BUCKET_NAME')
+                upload = upload_file_to_aws(file_name=audio_file_name, bucket=bucket_name)
 
-        image_file = request.data['image_file']
-        images = []
+                if upload is True:
+                    os.remove(audio_file_name)
 
-        if image_file:
-            file_format, img_str = image_file.split(';base64,')
-            ext = file_format.split('/')[-1]
-            img_file_name = f"{user_id}_{time.time()}_place_image.{ext}"
+            image_file = request.data['image_file']
+            images = []
 
-            with open(img_file_name, 'wb') as destination:
-                destination.write(base64.b64decode(img_str))
+            if image_file:
+                file_format, img_str = image_file.split(';base64,')
+                ext = file_format.split('/')[-1]
+                img_file_name = f"{user_id}_{time.time()}_place_image.{ext}"
 
-            bucket_name = os.getenv('AWS_AVATAR_IMAGE_BUCKET_NAME')
+                with open(img_file_name, 'wb') as destination:
+                    destination.write(base64.b64decode(img_str))
 
-            upload = upload_file_to_aws(file_name=img_file_name, bucket=bucket_name)
-            if upload is True:
-                os.remove(img_file_name)
-                images.append(img_file_name)
+                bucket_name = os.getenv('AWS_AVATAR_IMAGE_BUCKET_NAME')
 
-        place = Places.objects.create(name=data['name'], description=data['description'],
-                                      address=data['address'], longitude=data['longitude'],
-                                      latitude=data['latitude'], place_type_id=data['place_type'],
-                                      images=images, price_category_id=data['price_category'],
-                                      created_by_user_id=user_id, audio_url=audio_file_name)
+                upload = upload_file_to_aws(file_name=img_file_name, bucket=bucket_name)
+                if upload is True:
+                    os.remove(img_file_name)
+                    images.append(img_file_name)
 
-        return Response(data=PlaceSerializer(place).data, status=status.HTTP_201_CREATED)
+            place = Places.objects.create(name=data['name'], description=data['description'],
+                                          address=data['address'], longitude=data['longitude'],
+                                          latitude=data['latitude'], place_type_id=data['place_type'],
+                                          images=images, price_category_id=data['price_category'],
+                                          created_by_user_id=user_id, audio_url=audio_file_name)
+
+            return Response(data=PlaceSerializer(place).data, status=status.HTTP_201_CREATED)
+        else:
+            return Response({'error', 'Please verify your email address.'}, status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     def update(request, pk=None):
